🧹 chore(parametric): merge the two conversion modules and the stray promotion rules
**Two conversion modules, one body.** `register_conversions.py` and
`register_astropy.py` each registered a single `@conversion_method` to
`ParametricQuantity`, and both bodies were the same two lines: read the unit,
strip the value in it, rebuild. Merged into one module with a shared
`_to_parametric` helper.

**Promotion rules lived in two places.** `register_promotions.py` was a
9-line file holding two `add_promotion_rule` calls while a third sat at the
bottom of `parametric.py`. All three now sit together next to the class, and
the file (plus its side-effect import) is gone.

Verified the deleted modules' side effects survived rather than trusting a
green suite: `promote()` on Angle/PQ, StaticQuantity/PQ and Quantity/PQ all
still yield parametric pairs, `convert(astropy_quantity, PQ)` and
`convert(Quantity, PQ)` still resolve, and `Angle + PQ` / `PQ * Quantity`
still carry the dimension parameter.

Pure source reorganisation -- no dependency change, so this needs no astropy
floor bump. Aligning that floor with unxt's `>=7.1` is split out separately.
