+++ packages/unxts.parametric/src/unxts/parametric/__init__.py
@@ -11,10 +11,8 @@ __all__ = (
 # Import register modules for their dispatch/promotion side effects.
 from ._src import (  # noqa: F401
     register_api,
-    register_astropy,
     register_conversions,
     register_primitives,
-    register_promotions,
 )
 from ._src.base_parametric import AbstractParametricQuantity
 from ._src.config import config
