+++ packages/unxts.parametric/src/unxts/parametric/_src/register_conversions.py
@@ -1,13 +1,29 @@
-"""Conversions to ParametricQuantity (registered on import)."""
+"""Conversions to `ParametricQuantity` (registered on import).
+
+Both an `unxt` quantity and an `astropy.units.Quantity` convert the same way --
+read the unit, strip the value in it, rebuild -- so the two `plum` conversion
+methods share one body. ``astropy`` is a hard dependency of ``unxts.parametric``
+(its ``PhysicalType`` machinery requires it), so its conversion is always
+registered.
+"""
 
 __all__: tuple[str, ...] = ()
 
+from typing import Any
+
+from astropy.units import Quantity as AstropyQuantity
 from plum import conversion_method
 
 from unxts.api import unit_of, ustrip
 
 from .parametric import ParametricQuantity
 from unxt.quantity import AbstractQuantity
+
+
+def _to_parametric(q: Any, /) -> ParametricQuantity:
+    """Rebuild ``q`` as a `ParametricQuantity` in its own unit."""
+    u = unit_of(q)
+    return ParametricQuantity(ustrip(u, q), u)
 
 
 @conversion_method(type_from=AbstractQuantity, type_to=ParametricQuantity)  # type: ignore[arg-type]
@@ -36,5 +52,21 @@ def quantity_to_checked(q: AbstractQuantity, /) -> ParametricQuantity:
     """
     if isinstance(q, ParametricQuantity):
         return q
-    u = unit_of(q)
-    return ParametricQuantity(ustrip(u, q), u)
+    return _to_parametric(q)
+
+
+@conversion_method(type_from=AstropyQuantity, type_to=ParametricQuantity)  # type: ignore[arg-type]
+def astropy_quantity_to_parametric(q: AstropyQuantity, /) -> ParametricQuantity:
+    """Convert an `astropy.units.Quantity` to a `ParametricQuantity`.
+
+    Examples
+    --------
+    >>> from astropy.units import Quantity as AstropyQuantity
+    >>> from plum import convert
+    >>> from unxts.parametric import ParametricQuantity
+
+    >>> convert(AstropyQuantity(1.0, "cm"), ParametricQuantity)
+    ParametricQuantity(Array(1., dtype=float32), unit='cm')
+
+    """
+    return _to_parametric(q)
